use jsx and add success page
