--- src/components/Disclaimer.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const Disclaimer = () => {
-  return (
-    <div id="disclaimer" className="container-fluid text-center bg-warning text-dark">
-
-      <p>We and our partners use cookies to personalize your experience, to show you ads based on your interests, and for measurement and analytics purposes. By using our website and our services, you agree to our use of cookies as described in our Cookie Policy.</p>
-
-    </div>
-
-  )
-};
-
-export default Disclaimer;
